Fix for Issue #207 and farming tweaks (#208)
* Add files via upload

Sets cutting board to be both general food crafting and the avali kitchen recipes (may want to rename category to avoid collision with other mods?)

* Puts the food recipes in food subcategories

puts all but flour in "mains", flour recipe in "sides"

* Fix for recipe

recipe requires an item that no longer exists

* Balance adjustments to food

changes the pixel value and hunger values to closer to vanilla's food for similar items. Avali crops still have a shorter growth time, so the prices are lower. Hunger value still seems high to me, but is a lot lower than it was.

* Adds Avali food and crops to Terramart Shipments