Abandoning elevators, introducing railplatforms
Elevators are too single-use, rail platforms can go every which way, even diagonal and sideways and in a loop-de-loop!

You can also make custom-height elevators with it. 
(And Elevators kept breaking on my end. If anyone else wants to try to fix them be my guest, but for now it's rail platforms all the way. Probably going to add in aerogel rails based after the old elevator's middle bit)